+++ karchangame/src/main/java/mmud/rest/services/EventsBean.java
@@ -115,7 +115,7 @@ public class EventsBean
     Items items = new Items(new ItemsInterface()
     {
       @Override
-      public Item createItem(int itemdefnr)
+      public Item createItem(long itemdefnr)
       {
         return itemBean.createItem(itemdefnr);
       }
@@ -187,7 +187,7 @@ public class EventsBean
     Items items = new Items(new ItemsInterface()
     {
       @Override
-      public Item createItem(int itemdefnr)
+      public Item createItem(long itemdefnr)
       {
         return itemBean.createItem(itemdefnr);
       }
